refactor: move Map.groupBy into shims
Here moved our utility polyfill for Map.groupBy into shims module
we are using for all stuff we will replace with native methods.

Used it in marketplace to group by category.
