+++ apps/builder/app/shared/shim.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "@jest/globals";
-import { setDifference } from "./shim";
+import { mapGroupBy, setDifference } from "./shim";
 
 test("set difference", () => {
   // this set is bigger than other
@@ -9,5 +9,14 @@ test("set difference", () => {
   // this set is smaller than other
   expect(setDifference(new Set([1, 2, 3]), new Set([2, 3, 4, 5]))).toEqual(
     new Set([1])
+  );
+});
+
+test("groupBy", () => {
+  expect(mapGroupBy([1, 2, 3, 4, 5], (item) => item % 2)).toEqual(
+    new Map([
+      [0, [2, 4]],
+      [1, [1, 3, 5]],
+    ])
   );
 });
