+++ apps/builder/app/shared/shim.ts
@@ -15,3 +15,20 @@ export const setDifference = <Item>(current: Set<Item>, other: Set<Item>) => {
   }
   return result;
 };
+
+export const mapGroupBy = <Item, Key>(
+  array: Item[] | IterableIterator<Item>,
+  getKey: (item: Item) => Key
+) => {
+  const groups = new Map<Key, Item[]>();
+  for (const item of array) {
+    const key = getKey(item);
+    let group = groups.get(key);
+    if (group === undefined) {
+      group = [];
+      groups.set(key, group);
+    }
+    group.push(item);
+  }
+  return groups;
+};
